Scope modules to model implementation
This has no code change and only moves all the current implementation
to the `NervesTime.RTC.Abracon.IBO5` scope to allow for upcoming
changes supporting another Abracon model
