--- .tool-versions
@@ -1,1 +1,0 @@
-erlang 23.3.1
