+++ lib/nerves_time/rtc/abracon/ibo5/id.ex
@@ -1,4 +1,4 @@
-defmodule NervesTime.RTC.Abracon.ID do
+defmodule NervesTime.RTC.Abracon.IBO5.ID do
   @moduledoc false
 
   @doc """
