+++ test/nerves_time/rtc/abracon/id_test.exs
@@ -1,6 +1,6 @@
-defmodule NervesTime.RTC.Abracon.IDTest do
+defmodule NervesTime.RTC.Abracon.IBO5.IDTest do
   use ExUnit.Case
-  alias NervesTime.RTC.Abracon.ID
+  alias NervesTime.RTC.Abracon.IBO5.ID
 
   test "decodes ID" do
     assert ID.decode(<<24, 5, 19, 49, 70, 89, 48>>) ==
